task_backend/requirements.txt
amqp==5.1.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
async-timeout==4.0.3
auth0-python==4.4.0
billiard==4.1.0
black==23.7.0
celery==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.0
cryptography==41.0.3
defusedxml==0.7.1
Django==4.2.4
django-celery-beat==2.5.0
django-redis==5.3.0
django-timezone-field==5.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-extra-fields==3.7.0
drf-yasg==1.21.7
ecdsa==0.18.0
filetype==1.2.0
gunicorn==21.2.0
idna==3.4
inflection==0.5.1
kombu==5.3.1
load-dotenv==0.1.0
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.1
pathspec==0.11.2
Pillow==10.0.0
platformdirs==3.10.0
prompt-toolkit==3.0.39
psycopg2==2.9.6
psycopg2-binary==2.9.7
pyasn1==0.5.0
pycparser==2.21
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
redis==5.0.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
